ParseUser details to be removed
