+++ app/build.gradle
@@ -37,7 +37,7 @@ dependencies {
     compile 'de.hdodenhof:circleimageview:1.2.1'
     compile 'com.android.support:recyclerview-v7:21.0.0'
     compile 'com.android.support:appcompat-v7:22.2.0'
-    compile 'com.google.android.gms:play-services:7.5.0'
+    compile 'com.google.android.gms:play-services:7.3.0'
     compile 'com.parse.bolts:bolts-android:1.+'
     compile 'com.github.navasmdc:MaterialDesign:1.+@aar'
     compile ('com.nineoldandroids:library:2.4.0')
