+++ app/src/main/java/com/gogreen/greenmachine/util/Ifunction.java
@@ -1,0 +1,9 @@
+package com.gogreen.greenmachine.util;
+
+/**
+ * Created by arbkhan on 8/12/2015.
+ */
+//Interface for implementing function pointers
+public interface Ifunction {
+    void execute();
+}
